% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{get_step}
\alias{get_step}
\title{Returns the current time step of simulations.}
\usage{
get_step(simulations)
}
\arguments{
\item{simulations}{Either a single reference to a simulation object or a list
of such references.}
}
\value{
A numeric vector of the same length as \code{simulations}.
}
\description{
\code{get_step()} returns a numeric vector indicating the current time step of
each referenced simulation object. Those time steps usually (but not
necessarily) represent days.
}
