% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poseidr-package.R
\docType{package}
\name{poseidr-package}
\alias{poseidr}
\alias{poseidr-package}
\title{poseidr: The R interface to the POSEIDON agent-based model of fisheries}
\description{
The R interface to the POSEIDON agent-based model of fisheries.
}
\author{
\strong{Maintainer}: Nicolas Payette \email{nicolas.payette@ouce.ox.ac.uk} (\href{https://orcid.org/0000-0002-6628-4953}{ORCID})

}
\keyword{internal}
